pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Kiryl Liavonau <kliavonau@ibagroup.eu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
pydantic-settings = "^2.0.3"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.2"
httpx = "^0.25.0"
mypy = "^1.6.0"
black = "^23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
